pyproject.toml
[project]
requires-python = "~=3.11"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
pygame = "^2.5.2"
pyserial = "^3.5"
arducamdepthcamera = "^0.1.3"
opencv-python = "^4.9.0.80"
imutils = "^0.5.4"

[tool.poetry.group.dev.dependencies]
types-pyserial = "^3.5.0.20240311"

[tool.ruff.lint]
select = ["ALL"]
preview = true
ignore = ["S311", "CPY", "E501", "FA", "TD002", "TD003", "S301", "S403"]
